+++ packages/event-recurrence/src/event-recurrence-plugin.impl.ts
@@ -1,0 +1,94 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { EventRecurrencePlugin } from '@schedule-x/shared/src/interfaces/event-recurrence/event-recurrence-plugin.interface'
+import { CalendarAppSingleton } from '@schedule-x/shared/src'
+import { EventId } from '@schedule-x/shared/src/types/event-id'
+import { PluginName } from '@schedule-x/shared/src/enums/plugin-name.enum'
+import { CalendarEventInternal } from '@schedule-x/shared/src/interfaces/calendar/calendar-event.interface'
+import { DndUpdater } from './util/stateful/dnd-updater'
+import EventsFacade from '@schedule-x/shared/src/utils/stateful/events-facade/events-facade.interface'
+import { EventsFacadeImpl } from './util/stateful/events-facade'
+import { createRecurrencesForEvent } from './util/stateless/create-recurrences-for-event'
+import { ResizeUpdater } from './util/stateful/resize-updater'
+
+class EventRecurrencePluginImpl implements EventRecurrencePlugin {
+  name: string = PluginName.EventRecurrence
+  private $app: CalendarAppSingleton | null = null
+  init($app: CalendarAppSingleton): void {
+    this.$app = $app
+    this.createRecurrencesForEvents()
+  }
+
+  get eventsFacade(): EventsFacade {
+    if (!this.$app) throw new Error('Plugin not yet initialized')
+
+    return new EventsFacadeImpl(this.$app)
+  }
+
+  updateRecurrenceDND(
+    eventId: EventId,
+    oldEventStart: string,
+    newEventStart: string
+  ): void {
+    const { updatedEvent, recurrenceSet } = new DndUpdater(
+      this.$app as CalendarAppSingleton
+    ).update(eventId, oldEventStart, newEventStart)
+
+    this.$app!.calendarEvents.list.value = [
+      ...this.$app!.calendarEvents.list.value,
+      ...this.createRecurrencesForEvent(updatedEvent, recurrenceSet.getRrule()),
+    ]
+  }
+
+  updateRecurrenceOnResize(
+    eventId: EventId,
+    oldEventEnd: string,
+    newEventEnd: string
+  ): void {
+    const updatedEvent = new ResizeUpdater(
+      this.$app as CalendarAppSingleton
+    ).update(eventId, oldEventEnd, newEventEnd)
+
+    this.$app!.calendarEvents.list.value = [
+      ...this.$app!.calendarEvents.list.value,
+      ...this.createRecurrencesForEvent(
+        updatedEvent,
+        updatedEvent._getForeignProperties().rrule as string
+      ),
+    ]
+  }
+
+  private createRecurrencesForEvents() {
+    const recurrencesToCreate: CalendarEventInternal[] = []
+    const $app = this.$app as CalendarAppSingleton
+
+    $app.calendarEvents.list.value.forEach((event) => {
+      const rrule = event._getForeignProperties().rrule as string | undefined
+
+      if (rrule) {
+        recurrencesToCreate.push(
+          ...this.createRecurrencesForEvent(event, rrule)
+        )
+      }
+    })
+
+    $app.calendarEvents.list.value = [
+      ...this.$app!.calendarEvents.list.value,
+      ...recurrencesToCreate,
+    ]
+  }
+
+  private createRecurrencesForEvent(
+    calendarEvent: CalendarEventInternal,
+    rrule: string
+  ) {
+    return createRecurrencesForEvent(
+      this.$app as CalendarAppSingleton,
+      calendarEvent,
+      rrule
+    )
+  }
+}
+
+export const createEventRecurrencePlugin = (): EventRecurrencePlugin => {
+  return new EventRecurrencePluginImpl() as EventRecurrencePlugin
+}