+++ packages/event-recurrence/src/types/augmented-event.ts
@@ -1,0 +1,5 @@
+import { CalendarEventInternal } from '@schedule-x/shared/src/interfaces/calendar/calendar-event.interface'
+
+export interface AugmentedEvent extends CalendarEventInternal {
+  isCopy?: boolean
+}