+++ packages/shared/src/interfaces/event-recurrence/event-recurrence-plugin.interface.ts
@@ -1,10 +1,19 @@
 import PluginBase from '../plugin.interface'
 import { EventId } from '../../types/event-id'
+import EventsFacade from '../../utils/stateful/events-facade/events-facade.interface'
 
 export interface EventRecurrencePlugin extends PluginBase {
   updateRecurrenceDND(
     eventId: EventId,
-    newEventStart: string,
+    oldEventStart: string,
+    newEventStart: string
+  ): void
+
+  updateRecurrenceOnResize(
+    eventId: EventId,
+    oldEventEnd: string,
     newEventEnd: string
   ): void
+
+  eventsFacade: EventsFacade
 }