+++ packages/translations/src/index.ts
@@ -16,6 +16,8 @@ import { ruRU } from './locales/ru-RU'
 import { datePickerRuRU } from './locales/ru-RU/date-picker'
 import { koKR } from './locales/ko-KR'
 import { datePickerKoKR } from './locales/ko-KR/date-picker'
+import { frFR } from './locales/fr-FR'
+import { datePickerFrFR } from './locales/fr-FR/date-picker'
 import { daDK } from './locales/da-DK'
 import { datePickerDaDK } from './locales/da-DK/date-picker'
 
@@ -31,6 +33,7 @@ const translations = {
   jaJP,
   ruRU,
   koKR,
+  frFR,
   daDK,
 }
 
@@ -44,6 +47,7 @@ const datePickerTranslations = {
   jaJP: datePickerJaJP,
   ruRU: datePickerRuRU,
   koKR: datePickerKoKR,
+  frFR: datePickerFrFR,
   daDK: datePickerDaDK,
 }
 
@@ -60,5 +64,6 @@ export {
   jaJP,
   ruRU,
   koKR,
+  frFR,
   daDK,
 }