+++ packages/translations/src/locales/fr-FR/date-picker.ts
@@ -1,0 +1,10 @@
+import { DatePickerTranslations } from '../../types/date-picker.translations'
+
+export const datePickerFrFR: DatePickerTranslations = {
+  Date: 'Date',
+  'MM/DD/YYYY': 'JJ/MM/AAAA',
+  'Next month': 'Mois suivant',
+  'Previous month': 'Mois précédent',
+  'Choose Date': 'Choisir une date',
+  'Select View': 'Choisir la vue',
+}