+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/schedule-x/schedule-x/compare/v1.14.0...v1.15.0) (2024-02-06)
+
+### Features
+
+- update recurrent events on resize ([#225](https://github.com/schedule-x/schedule-x/issues/225)) ([dd522d4](https://github.com/schedule-x/schedule-x/commit/dd522d422c53a247d3a8661a15214d0349679922))
+
+# [1.14.0](https://github.com/schedule-x/schedule-x/compare/v1.13.0...v1.14.0) (2024-02-05)
+
+### Features
+
+- add french lang ([#222](https://github.com/schedule-x/schedule-x/issues/222)) ([ee5581a](https://github.com/schedule-x/schedule-x/commit/ee5581a2b209471be5499147a7ec9443971bfc21))
+
+# [1.13.0](https://github.com/schedule-x/schedule-x/compare/v1.12.2...v1.13.0) (2024-02-05)
+
+### Features
+
+- add events facade for recurrence plugin ([#220](https://github.com/schedule-x/schedule-x/issues/220)) ([dde1806](https://github.com/schedule-x/schedule-x/commit/dde180679d18f6ecf72a6616e824a9f9e321aa7a))
+- enable dragging a recurring events ([#219](https://github.com/schedule-x/schedule-x/issues/219)) ([35441e7](https://github.com/schedule-x/schedule-x/commit/35441e74546f4a9fe7c41e3f321bf6ecf1a4fb2b))
+- enable showing weekly events through rrule ([#218](https://github.com/schedule-x/schedule-x/issues/218)) ([6f33d0f](https://github.com/schedule-x/schedule-x/commit/6f33d0f8c29f35115502a744b3dc6816b75f9495))
+- recurrence library ([#208](https://github.com/schedule-x/schedule-x/issues/208)) ([f7488bd](https://github.com/schedule-x/schedule-x/commit/f7488bd89862e6a04d81452c067cfec6866a8ea6))
+
 ## [1.12.2](https://github.com/schedule-x/schedule-x/compare/v1.12.1...v1.12.2) (2024-02-02)
 
 ### Bug Fixes