+++ packages/resize/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/schedule-x/schedule-x/compare/v1.14.0...v1.15.0) (2024-02-06)
+
+### Features
+
+- update recurrent events on resize ([#225](https://github.com/schedule-x/schedule-x/issues/225)) ([dd522d4](https://github.com/schedule-x/schedule-x/commit/dd522d422c53a247d3a8661a15214d0349679922))
+
+# [1.14.0](https://github.com/schedule-x/schedule-x/compare/v1.13.0...v1.14.0) (2024-02-05)
+
+**Note:** Version bump only for package @schedule-x/resize
+
+# [1.13.0](https://github.com/schedule-x/schedule-x/compare/v1.12.2...v1.13.0) (2024-02-05)
+
+**Note:** Version bump only for package @schedule-x/resize
+
 ## [1.12.2](https://github.com/schedule-x/schedule-x/compare/v1.12.1...v1.12.2) (2024-02-02)
 
 **Note:** Version bump only for package @schedule-x/resize